Split preprocessing to separate class. Test strict face detection (box -> 160x160, [-1,1]) + Add better result.
